add check for admin only routes
