+++ litellm/proxy/proxy_config.yaml
@@ -12,5 +12,8 @@ litellm_settings:
 
 general_settings: 
  master_key: sk-1234 
- custom_auth: example_config_yaml.custom_auth_basic.user_api_key_auth
- allowed_routes: []
+
+ allowed_routes: ["/chat/completions"]
+ public_routes: ["LiteLLMRoutes.public_routes"]
+ admin_only_routes: ["/key/generate"]
+
